+++ integrationtests/src/test/java/io/kroxylicious/proxy/filter/ApiVersionsMarkingFilter.java
@@ -39,4 +39,5 @@ public class ApiVersionsMarkingFilter implements RequestFilter {
             return context.forwardRequest(header, request);
         });
     }
+
 }