+++ kroxylicious-additional-filters/kroxylicious-multitenant/src/main/java/io/kroxylicious/proxy/filter/multitenant/MultiTenantTransformationFilter.java
@@ -498,4 +498,5 @@ public class MultiTenantTransformationFilter
 
     public MultiTenantTransformationFilter() {
     }
+
 }