+++ kroxylicious-additional-filters/kroxylicious-schema-validation/src/main/java/io/kroxylicious/proxy/filter/schema/ProduceValidationFilter.java
@@ -187,4 +187,5 @@ public class ProduceValidationFilter implements ProduceRequestFilter, ProduceRes
             topicProduceResponse.partitionResponses().add(response);
         });
     }
+
 }