+++ kroxylicious-api/src/main/java/io/kroxylicious/proxy/clusternetworkaddressconfigprovider/ClusterNetworkAddressConfigProviderContributor.java
@@ -11,7 +11,8 @@ import io.kroxylicious.proxy.service.Contributor;
 
 /**
  * ClusterNetworkAddressConfigProviderContributor is a pluggable source of network address information.
+ * @param <B> config type
  * @see Contributor
  */
-public interface ClusterNetworkAddressConfigProviderContributor extends Contributor<ClusterNetworkAddressConfigProvider, Context> {
+public interface ClusterNetworkAddressConfigProviderContributor<B> extends Contributor<ClusterNetworkAddressConfigProvider, B, Context<B>> {
 }