+++ kroxylicious-api/src/main/java/io/kroxylicious/proxy/service/ConfigurationDefinition.java
@@ -6,10 +6,8 @@
 
 package io.kroxylicious.proxy.service;
 
-import io.kroxylicious.proxy.config.BaseConfig;
-
 /**
- * Defines the details of how Contributions interact with configuration for {@link BaseContributor}'s purposes.
+ * Defines the details of how Contributions interact with configuration.
  * <p>
  * configurationRequired allows filter authors to tell Kroxylicious if the filter must be supplied with a config object.
  * If the filter uses configuration to supply optional properties, or it can provide sensible default values for the configuration then it should be set to false.
@@ -17,4 +15,4 @@ import io.kroxylicious.proxy.config.BaseConfig;
  * @param configurationType defines the expected class for configuration objects
  * @param configurationRequired {@code true} if the contribution requires a non-null configuration object. {@code false} if the contribution can tolerate a null configuration object
  */
-public record ConfigurationDefinition(Class<? extends BaseConfig> configurationType, boolean configurationRequired) {}
+public record ConfigurationDefinition(Class<?> configurationType, boolean configurationRequired) {}