+++ kroxylicious/src/main/java/io/kroxylicious/proxy/config/FilterConfigTypeIdResolver.java
@@ -5,15 +5,12 @@
  */
 package io.kroxylicious.proxy.config;
 
-import java.io.IOException;
-
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.databind.DatabindContext;
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.jsontype.impl.TypeIdResolverBase;
 
-import io.kroxylicious.proxy.filter.FilterContributor;
-import io.kroxylicious.proxy.service.ContributionManager;
+import io.kroxylicious.proxy.service.FilterFactoryManager;
 
 public class FilterConfigTypeIdResolver extends TypeIdResolverBase {
     private JavaType superType;
@@ -39,9 +36,8 @@ public class FilterConfigTypeIdResolver extends TypeIdResolverBase {
     }
 
     @Override
-    public JavaType typeFromId(DatabindContext context, String id) throws IOException {
-        Class<?> subType = ContributionManager.INSTANCE.getDefinition(FilterContributor.class, id)
-                .configurationType();
+    public JavaType typeFromId(DatabindContext context, String id) {
+        Class<?> subType = FilterFactoryManager.INSTANCE.getConfigType(id);
         return context.constructSpecializedType(superType, subType);
     }
 }