+++ kroxylicious/src/main/java/io/kroxylicious/proxy/config/FilterDefinition.java
@@ -13,12 +13,11 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import com.fasterxml.jackson.databind.annotation.JsonTypeIdResolver;
 
-import io.kroxylicious.proxy.filter.FilterContributor;
-import io.kroxylicious.proxy.service.ConfigurationDefinition;
-import io.kroxylicious.proxy.service.ContributionManager;
+import io.kroxylicious.proxy.service.FilterFactoryManager;
 
 public record FilterDefinition(@JsonProperty(required = true) String type,
-                               @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.EXTERNAL_PROPERTY, property = "type") @JsonTypeIdResolver(FilterConfigTypeIdResolver.class) BaseConfig config) {
+                               @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.EXTERNAL_PROPERTY, property = "type") @JsonTypeIdResolver(FilterConfigTypeIdResolver.class) Object config) {
+
     @JsonCreator
     public FilterDefinition {
         Objects.requireNonNull(type);
@@ -26,9 +25,6 @@ public record FilterDefinition(@JsonProperty(required = true) String type,
 
     @JsonIgnore
     public boolean isDefinitionValid() {
-        // We could consider allowing Filter authors to supply a config validation function `config -> config.property != null` via the configurationDefinition
-        // We would then be able to apply that validation here.
-        final ConfigurationDefinition configurationDefinition = ContributionManager.INSTANCE.getDefinition(FilterContributor.class, type);
-        return config != null || !configurationDefinition.configurationRequired();
+        return FilterFactoryManager.INSTANCE.validateConfig(type, config);
     }
 }