+++ kroxylicious/src/main/java/io/kroxylicious/proxy/internal/clusternetworkaddressconfigprovider/PortPerBrokerClusterNetworkAddressConfigProvider.java
@@ -17,7 +17,6 @@ import java.util.stream.IntStream;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-import io.kroxylicious.proxy.config.BaseConfig;
 import io.kroxylicious.proxy.service.ClusterNetworkAddressConfigProvider;
 import io.kroxylicious.proxy.service.HostPort;
 
@@ -101,7 +100,7 @@ public class PortPerBrokerClusterNetworkAddressConfigProvider implements Cluster
     /**
      * Creates the configuration for this provider.
      */
-    public static class PortPerBrokerClusterNetworkAddressConfigProviderConfig extends BaseConfig {
+    public static class PortPerBrokerClusterNetworkAddressConfigProviderConfig {
         private final HostPort bootstrapAddress;
         private final String brokerAddressPattern;
         private final int brokerStartPort;
@@ -146,4 +145,5 @@ public class PortPerBrokerClusterNetworkAddressConfigProvider implements Cluster
             return bootstrapAddress;
         }
     }
+
 }