+++ kroxylicious/src/main/java/io/kroxylicious/proxy/internal/filter/FetchResponseTransformationFilter.java
@@ -29,7 +29,6 @@ import org.apache.kafka.common.record.TimestampType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.kroxylicious.proxy.config.BaseConfig;
 import io.kroxylicious.proxy.filter.FetchResponseFilter;
 import io.kroxylicious.proxy.filter.FilterContext;
 import io.kroxylicious.proxy.filter.ResponseFilterResult;
@@ -40,7 +39,7 @@ import io.kroxylicious.proxy.internal.util.MemoryRecordsHelper;
  */
 public class FetchResponseTransformationFilter implements FetchResponseFilter {
 
-    public static class FetchResponseTransformationConfig extends BaseConfig {
+    public static class FetchResponseTransformationConfig {
 
         private final String transformation;
 
@@ -126,4 +125,5 @@ public class FetchResponseTransformationFilter implements FetchResponseFilter {
             }
         }
     }
+
 }